Created separate folder for configuration files Updated imports accordingly
