+++ configuration_modules/config_installed_sensors.py
