+++ http_server/flask_blueprints/sensor_control_files/sensor_control.py
@@ -23,8 +23,8 @@ from operations_modules import app_cached_variables
 from operations_modules import app_generic_functions
 from operations_modules import app_config_access
 from operations_modules import network_wifi
-from operations_modules.config_installed_sensors import CreateInstalledSensorsConfiguration
-from operations_modules.config_trigger_variances import CreateTriggerVariancesConfiguration
+from configuration_modules.config_installed_sensors import CreateInstalledSensorsConfiguration
+from configuration_modules.config_trigger_variances import CreateTriggerVariancesConfiguration
 from http_server.server_http_auth import auth
 from http_server.server_http_generic_functions import message_and_return, get_sensor_control_report
 from http_server.flask_blueprints.sensor_control_files.sensor_control_functions import \
