+++ test_http_server.py
@@ -23,13 +23,13 @@ from operations_modules.app_generic_functions import get_http_sensor_reading, ht
 from operations_modules import app_cached_variables
 from operations_modules.app_cached_variables import no_sensor_present, command_data_separator, \
     CreateNetworkGetCommands, CreateNetworkSetCommands
-from operations_modules.config_primary import CreatePrimaryConfiguration
-from operations_modules.config_installed_sensors import CreateInstalledSensorsConfiguration
-from operations_modules.config_trigger_variances import CreateTriggerVariancesConfiguration
-from operations_modules.config_sensor_control import CreateSensorControlConfiguration
-from operations_modules.config_weather_underground import CreateWeatherUndergroundConfiguration
-from operations_modules.config_luftdaten import CreateLuftdatenConfiguration
-from operations_modules.config_open_sense_map import CreateOpenSenseMapConfiguration
+from configuration_modules.config_primary import CreatePrimaryConfiguration
+from configuration_modules.config_installed_sensors import CreateInstalledSensorsConfiguration
+from configuration_modules.config_trigger_variances import CreateTriggerVariancesConfiguration
+from configuration_modules.config_sensor_control import CreateSensorControlConfiguration
+from configuration_modules.config_weather_underground import CreateWeatherUndergroundConfiguration
+from configuration_modules.config_luftdaten import CreateLuftdatenConfiguration
+from configuration_modules.config_open_sense_map import CreateOpenSenseMapConfiguration
 
 primary_config_test = CreatePrimaryConfiguration(load_from_file=False)
 installed_sensors_config_test = CreateInstalledSensorsConfiguration(load_from_file=False)
